+++ public/controllers/agent/components/register-agent.js
@@ -78,6 +78,7 @@ import {
 } from './register-agent-service';
 import './register-agent.scss';
 import { PrincipalButtonGroup } from './wz-accordion';
+import RegisterAgentButtonGroup from '../register-agent/register-agent-button-group';
 import '../../../styles/common.scss';
 
 export const RegisterAgent = withErrorBoundary(
@@ -1635,31 +1636,6 @@ apk add wazuh-agent=${this.state.wazuhVersion}-r1`,
         },
       ];
 
-      const buttonGroup = (legend, options, idSelected, onChange) => {
-        return (
-          <EuiButtonGroup
-            color='primary'
-            legend={legend}
-            options={options}
-            idSelected={idSelected}
-            onChange={onChange}
-            className={'wz-flex'}
-          />
-        );
-      };
-
-      const selectedVersionMac = (legend, options, idSelected, onChange) => {
-        return (
-          <EuiButtonGroup
-            color='primary'
-            legend={legend}
-            options={options}
-            idSelected={idSelected}
-            onChange={onChange}
-          />
-        );
-      };
-
       const onChangeServerAddress = async selectedNodes => {
         if (selectedNodes.length === 0) {
           this.setState({
@@ -1715,12 +1691,11 @@ apk add wazuh-agent=${this.state.wazuhVersion}-r1`,
           ? [
               {
                 title: 'Choose the version',
-                children: buttonGroup(
-                  'Choose the version',
-                  versionButtonsRedHat,
-                  this.state.selectedVersion,
-                  version => this.setVersion(version),
-                ),
+                children: <RegisterAgentButtonGroup
+                  legend='Choose the version'
+                  options={versionButtonsRedHat}
+                  idSelected={this.state.selectedVersion}
+                  onChange={version => this.setVersion(version)} />,
               },
             ]
           : []),
@@ -1728,12 +1703,11 @@ apk add wazuh-agent=${this.state.wazuhVersion}-r1`,
           ? [
               {
                 title: 'Choose the version',
-                children: buttonGroup(
-                  'Choose the version',
-                  versionButtonsOracleLinux,
-                  this.state.selectedVersion,
-                  version => this.setVersion(version),
-                ),
+                children: <RegisterAgentButtonGroup
+                  legend='Choose the version'
+                  options={versionButtonsOracleLinux}
+                  idSelected={this.state.selectedVersion}
+                  onChange={version => this.setVersion(version)} />,
               },
             ]
           : []),
@@ -1741,12 +1715,11 @@ apk add wazuh-agent=${this.state.wazuhVersion}-r1`,
           ? [
               {
                 title: 'Choose the version',
-                children: buttonGroup(
-                  'Choose the version',
-                  versionButtonsRaspbian,
-                  this.state.selectedVersion,
-                  version => this.setVersion(version),
-                ),
+                children: <RegisterAgentButtonGroup
+                  legend='Choose the version'
+                  options={versionButtonsRaspbian}
+                  idSelected={this.state.selectedVersion}
+                  onChange={version => this.setVersion(version)} />,
               },
             ]
           : []),
@@ -1754,12 +1727,11 @@ apk add wazuh-agent=${this.state.wazuhVersion}-r1`,
           ? [
               {
                 title: 'Choose the version',
-                children: buttonGroup(
-                  'Choose the version',
-                  versionButtonAmazonLinux,
-                  this.state.selectedVersion,
-                  version => this.setVersion(version),
-                ),
+                children: <RegisterAgentButtonGroup
+                  legend='Choose the version'
+                  options={versionButtonAmazonLinux}
+                  idSelected={this.state.selectedVersion}
+                  onChange={version => this.setVersion(version)} />,
               },
             ]
           : []),
@@ -1767,12 +1739,11 @@ apk add wazuh-agent=${this.state.wazuhVersion}-r1`,
           ? [
               {
                 title: 'Choose the version',
-                children: buttonGroup(
-                  'Choose the version',
-                  versionButtonsCentos,
-                  this.state.selectedVersion,
-                  version => this.setVersion(version),
-                ),
+                children: <RegisterAgentButtonGroup
+                  legend='Choose the version'
+                  options={versionButtonsCentos}
+                  idSelected={this.state.selectedVersion}
+                  onChange={version => this.setVersion(version)} />,
               },
             ]
           : []),
@@ -1780,12 +1751,11 @@ apk add wazuh-agent=${this.state.wazuhVersion}-r1`,
           ? [
               {
                 title: 'Choose the version',
-                children: buttonGroup(
-                  'Choose the version',
-                  versionButtonFedora,
-                  this.state.selectedVersion,
-                  version => this.setVersion(version),
-                ),
+                children: <RegisterAgentButtonGroup
+                  legend='Choose the version'
+                  options={versionButtonFedora}
+                  idSelected={this.state.selectedVersion}
+                  onChange={version => this.setVersion(version)} />,
               },
             ]
           : []),
@@ -1793,12 +1763,11 @@ apk add wazuh-agent=${this.state.wazuhVersion}-r1`,
           ? [
               {
                 title: 'Choose the version',
-                children: buttonGroup(
-                  'Choose the version',
-                  versionButtonsDebian,
-                  this.state.selectedVersion,
-                  version => this.setVersion(version),
-                ),
+                children: <RegisterAgentButtonGroup
+                  legend='Choose the version'
+                  options={versionButtonsDebian}
+                  idSelected={this.state.selectedVersion}
+                  onChange={version => this.setVersion(version)} />,
               },
             ]
           : []),
@@ -1806,12 +1775,11 @@ apk add wazuh-agent=${this.state.wazuhVersion}-r1`,
           ? [
               {
                 title: 'Choose the version',
-                children: buttonGroup(
-                  'Choose the version',
-                  versionButtonsUbuntu,
-                  this.state.selectedVersion,
-                  version => this.setVersion(version),
-                ),
+                children: <RegisterAgentButtonGroup
+                  legend='Choose the version'
+                  options={versionButtonsUbuntu}
+                  idSelected={this.state.selectedVersion}
+                  onChange={version => this.setVersion(version)} />,
               },
             ]
           : []),
@@ -1819,12 +1787,11 @@ apk add wazuh-agent=${this.state.wazuhVersion}-r1`,
           ? [
               {
                 title: 'Choose the version',
-                children: buttonGroup(
-                  'Choose the version',
-                  versionButtonsWindows,
-                  this.state.selectedVersion,
-                  version => this.setVersion(version),
-                ),
+                children: <RegisterAgentButtonGroup
+                  legend='Choose the version'
+                  options={versionButtonsWindows}
+                  idSelected={this.state.selectedVersion}
+                  onChange={version => this.setVersion(version)} />,
               },
             ]
           : []),
@@ -1832,12 +1799,11 @@ apk add wazuh-agent=${this.state.wazuhVersion}-r1`,
           ? [
               {
                 title: 'Choose the version',
-                children: selectedVersionMac(
-                  'Choose the version',
-                  versionButtonsMacOS,
-                  this.state.selectedVersion,
-                  version => this.setVersion(version),
-                ),
+                children: <RegisterAgentButtonGroup
+                  legend='Choose the version'
+                  options={versionButtonsMacOS}
+                  idSelected={this.state.selectedVersion}
+                  onChange={version => this.setVersion(version)} />,
               },
             ]
           : []),
@@ -1845,12 +1811,11 @@ apk add wazuh-agent=${this.state.wazuhVersion}-r1`,
           ? [
               {
                 title: 'Choose the version',
-                children: selectedVersionMac(
-                  'Choose the version',
-                  versionButtonsSuse,
-                  this.state.selectedVersion,
-                  version => this.setVersion(version),
-                ),
+                children: <RegisterAgentButtonGroup
+                  legend='Choose the version'
+                  options={versionButtonsSuse}
+                  idSelected={this.state.selectedVersion}
+                  onChange={version => this.setVersion(version)} />,
               },
             ]
           : []),
@@ -1858,12 +1823,11 @@ apk add wazuh-agent=${this.state.wazuhVersion}-r1`,
           ? [
               {
                 title: 'Choose the version',
-                children: buttonGroup(
-                  'Choose the version',
-                  versionButtonsOpenSuse,
-                  this.state.selectedVersion,
-                  version => this.setVersion(version),
-                ),
+                children: <RegisterAgentButtonGroup
+                  legend='Choose the version'
+                  options={versionButtonsOpenSuse}
+                  idSelected={this.state.selectedVersion}
+                  onChange={version => this.setVersion(version)} />,
               },
             ]
           : []),
@@ -1871,12 +1835,11 @@ apk add wazuh-agent=${this.state.wazuhVersion}-r1`,
           ? [
               {
                 title: 'Choose the version',
-                children: buttonGroup(
-                  'Choose the version',
-                  versionButtonsSolaris,
-                  this.state.selectedVersion,
-                  version => this.setVersion(version),
-                ),
+                children: <RegisterAgentButtonGroup
+                  legend='Choose the version'
+                  options={versionButtonsSolaris}
+                  idSelected={this.state.selectedVersion}
+                  onChange={version => this.setVersion(version)} />,
               },
             ]
           : []),
@@ -1884,12 +1847,11 @@ apk add wazuh-agent=${this.state.wazuhVersion}-r1`,
           ? [
               {
                 title: 'Choose the version',
-                children: buttonGroup(
-                  'Choose the version',
-                  versionButtonsAix,
-                  this.state.selectedVersion,
-                  version => this.setVersion(version),
-                ),
+                children: <RegisterAgentButtonGroup
+                  legend='Choose the version'
+                  options={versionButtonsAix}
+                  idSelected={this.state.selectedVersion}
+                  onChange={version => this.setVersion(version)} />,
               },
             ]
           : []),
@@ -1897,12 +1859,11 @@ apk add wazuh-agent=${this.state.wazuhVersion}-r1`,
           ? [
               {
                 title: 'Choose the version',
-                children: buttonGroup(
-                  'Choose the version',
-                  versionButtonsHPUX,
-                  this.state.selectedVersion,
-                  version => this.setVersion(version),
-                ),
+                children: <RegisterAgentButtonGroup
+                  legend='Choose the version'
+                  options={versionButtonsHPUX}
+                  idSelected={this.state.selectedVersion}
+                  onChange={version => this.setVersion(version)} />,
               },
             ]
           : []),
@@ -1910,12 +1871,11 @@ apk add wazuh-agent=${this.state.wazuhVersion}-r1`,
           ? [
               {
                 title: 'Choose the version',
-                children: buttonGroup(
-                  'Choose the version',
-                  versionButtonAlpine,
-                  this.state.selectedVersion,
-                  version => this.setVersion(version),
-                ),
+                children: <RegisterAgentButtonGroup
+                  legend='Choose the version'
+                  options={versionButtonAlpine}
+                  idSelected={this.state.selectedVersion}
+                  onChange={version => this.setVersion(version)} />,
               },
             ]
           : []),
@@ -1926,12 +1886,11 @@ apk add wazuh-agent=${this.state.wazuhVersion}-r1`,
           ? [
               {
                 title: 'Choose the architecture',
-                children: buttonGroup(
-                  'Choose the architecture',
-                  architecturei386Andx86_64,
-                  this.state.selectedArchitecture,
-                  architecture => this.setArchitecture(architecture),
-                ),
+                children: <RegisterAgentButtonGroup 
+                  legend='Choose the architecture'
+                  options={architecturei386Andx86_64}
+                  idSelected={this.state.selectedArchitecture}
+                  onChange={architecture => this.setArchitecture(architecture)} />
               },
             ]
           : []),
@@ -1939,12 +1898,11 @@ apk add wazuh-agent=${this.state.wazuhVersion}-r1`,
           ? [
               {
                 title: 'Choose the architecture',
-                children: buttonGroup(
-                  'Choose the architecture',
-                  architectureButtonsWithPPC64LE,
-                  this.state.selectedArchitecture,
-                  architecture => this.setArchitecture(architecture),
-                ),
+                children: <RegisterAgentButtonGroup 
+                  legend='Choose the architecture'
+                  options={architectureButtonsWithPPC64LE}
+                  idSelected={this.state.selectedArchitecture}
+                  onChange={architecture => this.setArchitecture(architecture)} />
               },
             ]
           : []),
@@ -1952,12 +1910,11 @@ apk add wazuh-agent=${this.state.wazuhVersion}-r1`,
           ? [
               {
                 title: 'Choose the architecture',
-                children: buttonGroup(
-                  'Choose the architecture',
-                  architectureButtonsWithPPC64LEAlpine,
-                  this.state.selectedArchitecture,
-                  architecture => this.setArchitecture(architecture),
-                ),
+                children: <RegisterAgentButtonGroup
+                  legend='Choose the architecture'
+                  options={architectureButtonsWithPPC64LEAlpine}
+                  idSelected={this.state.selectedArchitecture}
+                  onChange={architecture => this.setArchitecture(architecture)} />
               },
             ]
           : []),
@@ -1974,12 +1931,11 @@ apk add wazuh-agent=${this.state.wazuhVersion}-r1`,
           ? [
               {
                 title: 'Choose the architecture',
-                children: buttonGroup(
-                  'Choose the architecture',
-                  architectureButtons,
-                  this.state.selectedArchitecture,
-                  architecture => this.setArchitecture(architecture),
-                ),
+                children: <RegisterAgentButtonGroup
+                  legend='Choose the architecture'
+                  options={architectureButtons}
+                  idSelected={this.state.selectedArchitecture}
+                  onChange={architecture => this.setArchitecture(architecture)} />
               },
             ]
           : []),
@@ -1992,12 +1948,11 @@ apk add wazuh-agent=${this.state.wazuhVersion}-r1`,
           ? [
               {
                 title: 'Choose the architecture',
-                children: buttonGroup(
-                  'Choose the architecture',
-                  architectureButtonsWithPPC64LE,
-                  this.state.selectedArchitecture,
-                  architecture => this.setArchitecture(architecture),
-                ),
+                children: <RegisterAgentButtonGroup
+                  legend='Choose the architecture'
+                  options={architectureButtonsWithPPC64LE}
+                  idSelected={this.state.selectedArchitecture}
+                  onChange={architecture => this.setArchitecture(architecture)} />
               },
             ]
           : []),
@@ -2007,12 +1962,11 @@ apk add wazuh-agent=${this.state.wazuhVersion}-r1`,
           ? [
               {
                 title: 'Choose the architecture',
-                children: buttonGroup(
-                  'Choose the architecture',
-                  architectureButtonsi386,
-                  this.state.selectedArchitecture,
-                  architecture => this.setArchitecture(architecture),
-                ),
+                children: <RegisterAgentButtonGroup
+                  legend='Choose the architecture'
+                  options={architectureButtonsi386}
+                  idSelected={this.state.selectedArchitecture}
+                  onChange={architecture => this.setArchitecture(architecture)} />
               },
             ]
           : []),
@@ -2020,12 +1974,11 @@ apk add wazuh-agent=${this.state.wazuhVersion}-r1`,
           ? [
               {
                 title: 'Choose the architecture',
-                children: buttonGroup(
-                  'Choose the architecture',
-                  architectureButtonsMacos,
-                  this.state.selectedArchitecture,
-                  architecture => this.setArchitecture(architecture),
-                ),
+                children: <RegisterAgentButtonGroup
+                  legend='Choose the architecture'
+                  options={architectureButtonsMacos}
+                  idSelected={this.state.selectedArchitecture}
+                  onChange={architecture => this.setArchitecture(architecture)} />
               },
             ]
           : []),
@@ -2034,12 +1987,11 @@ apk add wazuh-agent=${this.state.wazuhVersion}-r1`,
           ? [
               {
                 title: 'Choose the architecture',
-                children: buttonGroup(
-                  'Choose the architecture',
-                  architectureButtonsSolaris,
-                  this.state.selectedArchitecture,
-                  architecture => this.setArchitecture(architecture),
-                ),
+                children: <RegisterAgentButtonGroup
+                  legend='Choose the architecture'
+                  options={architectureButtonsSolaris}
+                  idSelected={this.state.selectedArchitecture}
+                  onChange={architecture => this.setArchitecture(architecture)} />
               },
             ]
           : []),
@@ -2047,12 +1999,11 @@ apk add wazuh-agent=${this.state.wazuhVersion}-r1`,
           ? [
               {
                 title: 'Choose the architecture',
-                children: buttonGroup(
-                  'Choose the architecture',
-                  architectureButtonsAix,
-                  this.state.selectedArchitecture,
-                  architecture => this.setArchitecture(architecture),
-                ),
+                children: <RegisterAgentButtonGroup
+                  legend='Choose the architecture'
+                  options={architectureButtonsAix}
+                  idSelected={this.state.selectedArchitecture}
+                  onChange={architecture => this.setArchitecture(architecture)} />
               },
             ]
           : []),
@@ -2060,12 +2011,11 @@ apk add wazuh-agent=${this.state.wazuhVersion}-r1`,
           ? [
               {
                 title: 'Choose the architecture',
-                children: buttonGroup(
-                  'Choose the architecture',
-                  architectureButtonsHpUx,
-                  this.state.selectedArchitecture,
-                  architecture => this.setArchitecture(architecture),
-                ),
+                children: <RegisterAgentButtonGroup
+                  legend='Choose the architecture'
+                  options={architectureButtonsHpUx}
+                  idSelected={this.state.selectedArchitecture}
+                  onChange={architecture => this.setArchitecture(architecture)} />
               },
             ]
           : []),
@@ -2181,7 +2131,6 @@ apk add wazuh-agent=${this.state.wazuhVersion}-r1`,
                       this.state.selectedVersion == 'centos7' ||
                       this.state.selectedVersion == 'suse11' ||
                       this.state.selectedVersion == 'suse12' ||
-                      this.state.selectedVersion == 'oraclelinux5' ||
                       this.state.selectedVersion == 'amazonlinux2' ||
                       this.state.selectedVersion == '22' ||
                       this.state.selectedVersion == 'debian8' ||